% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mEBA_Rfunctions.R
\name{meba.simdata}
\alias{meba.simdata}
\title{function to simulate meba data for 5 different settings}
\usage{
meba.simdata(t)
}
\arguments{
\item{t}{restOfOwl}
}
\value{
restOfOwl
}
\description{
function to simulate meba data for 5 different settings
}
